tools/docs/requirements.txt
python-docx>=1.2.0
PyYAML>=6.0.0
openpyxl>=3.1.0
pillow>=10.0.0
python-pptx>=0.6.23
